pyproject.toml
[tool.poetry]
name = "olist-data"
version = "0.1.0"
description = ""
authors = ["Claudio Pontes <claudiohpontes@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
kaggle = "^1.6.17"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
dbt-core = "^1.8.5"
dbt-postgres = "^1.8.2"
duckdb = "^1.0.0"
pytest = "^8.3.2"
pydantic = "^2.8.2"
taskipy = "^1.13.0"
black = "^24.8.0"
isort = "^5.13.2"
boto3 = "^1.35.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
format = """
isort .
black .
sqlfluff fix dbt_dados_abertos_camara/models
"""
kill = "kill -9 $(lsof -t -i :8000)"
test = "pytest -v"
